--- logs/log_2024-04-22_17-35-23.log
@@ -1,13 +1,0 @@
-
------
-Test: test_krasnodar_220_volt_api.py::TestTakeCategoryAndProductList::test_take_category_and_product_list (call)
-Time: 2024-04-22 17:35:23.668267
-Request method: POST
-Request URL: https://krasnodar.220-volt.ru/tracker_advcake/
-
-Response code: 200
-Response text: {"currentCategory":{"name":"Сад и дача","id":339},"pageType":3,"products":[{"id":"735229","name":"Бензопила HYUNDAI Х 5320"},{"id":"735227","name":"Бензопила HYUNDAI Х 3916"},{"id":"735228","name":"Бензопила HYUNDAI Х 4118"},{"id":"697556","name":"Бензопила HYUNDAI X 520"},{"name":"Бензопила HYUNDAI X 420","id":"697555"},{"name":"Бензопила HYUNDAI X 370","id":"697554"}],"user":{"email":"fbdb1d1b18aa6c08324b7d64b71fb76370690e1d"},"basketProducts":[]}
-Response headers: {'Server': 'nginx', 'Date': 'Mon, 22 Apr 2024 14:35:25 GMT', 'Content-Type': 'application/json; charset=windows-1251', 'Transfer-Encoding': 'chunked', 'Connection': 'keep-alive', 'Vary': 'Accept-Encoding', 'Set-Cookie': 'chkuidsrv=35e2ca3a69fa466aadf3bed420a32884; expires=Mon, 22-Apr-24 15:35:25 GMT; max-age=3600; domain=.220-volt.ru; path=/, site_version=desktop; expires=Fri, 21 Jun 2024 14:35:25 GMT; path=/; domain=.220-volt.ru, site_version=desktop; expires=Fri, 21 Jun 2024 14:35:25 GMT; path=/; domain=.220-volt.ru, telemetryToken=A53WpsneRNDlDxriGk0KwPEIiLkN0vTmeIVx71dTKAUx; expires=Sat, 21 Apr 2029 14:35:25 GMT; path=/; domain=.220-volt.ru, backendStart=1713796525121; expires=Sat, 21 Apr 2029 14:35:25 GMT; path=/; domain=.220-volt.ru, backendEnd=1713796525215; expires=Sat, 21 Apr 2029 14:35:25 GMT; path=/; domain=.220-volt.ru, system_id=4f3329cdc920abf850bfb05963ca7bc6z1713796525; path=/;Max-Age=1800;Domain=.220-volt.ru', 'Expires': 'Thu, 01 Jan 1970 00:00:00 GMT', 'Pragma': 'no-cache', 'Access-Control-Allow-Methods': 'OPTIONS, GET, POST, PUT, HEAD, PATCH, DELETE, OPTIONS', 'Cache-Control': 'no-store, no-cache, must-revalidate, max-age=0', 'X-Elapsed': '0.0939', 'Access-Control-Allow-Headers': 'Content-Type', 'Access-Control-Allow-Origin': 'https://*220-volt.ru', 'Last-Modified': 'Mon, 22 Apr 2024 14:35:25 GMT', 'RID': 'ZiZ1rct1tGoFd3csAeTz8QAAABY', 'Content-Encoding': 'gzip'}
-Response cookies: {'chkuidsrv': '35e2ca3a69fa466aadf3bed420a32884', 'site_version': 'desktop', 'telemetryToken': 'A53WpsneRNDlDxriGk0KwPEIiLkN0vTmeIVx71dTKAUx', 'backendStart': '1713796525121', 'backendEnd': '1713796525215', 'system_id': '4f3329cdc920abf850bfb05963ca7bc6z1713796525'}
-
------
