+++ logs/log_2024-04-22_22-42-17.log
@@ -1,0 +1,13 @@
+
+-----
+Test: test_krasnodar_220_volt_api.py::TestTakeCategoryAndProductList::test_validate_data (call)
+Time: 2024-04-22 22:42:18.204918
+Request method: POST
+Request URL: https://krasnodar.220-volt.ru/tracker_advcake/
+
+Response code: 200
+Response text: {"user":{"email":"fbdb1d1b18aa6c08324b7d64b71fb76370690e1d"},"basketProducts":[],"currentCategory":{"id":339,"name":"Сад и дача"},"pageType":3,"products":[{"id":"735229","name":"Бензопила HYUNDAI Х 5320"},{"id":"735227","name":"Бензопила HYUNDAI Х 3916"},{"id":"735228","name":"Бензопила HYUNDAI Х 4118"},{"name":"Бензопила HYUNDAI X 520","id":"697556"},{"id":"697555","name":"Бензопила HYUNDAI X 420"},{"id":"697554","name":"Бензопила HYUNDAI X 370"}]}
+Response headers: {'Server': 'nginx', 'Date': 'Mon, 22 Apr 2024 19:42:19 GMT', 'Content-Type': 'application/json; charset=windows-1251', 'Transfer-Encoding': 'chunked', 'Connection': 'keep-alive', 'Vary': 'Accept-Encoding', 'Set-Cookie': 'chkuidsrv=35e2ca3a69fa466aadf3bed420a32884; expires=Mon, 22-Apr-24 20:42:19 GMT; max-age=3600; domain=.220-volt.ru; path=/, site_version=desktop; expires=Fri, 21 Jun 2024 19:42:19 GMT; path=/; domain=.220-volt.ru, site_version=desktop; expires=Fri, 21 Jun 2024 19:42:19 GMT; path=/; domain=.220-volt.ru, telemetryToken=A53WpsneRNDlDxriGk0KwP5wDuY6VIbuRi2EimNKzvuP; expires=Sat, 21 Apr 2029 19:42:19 GMT; path=/; domain=.220-volt.ru, backendStart=1713814939412; expires=Sat, 21 Apr 2029 19:42:19 GMT; path=/; domain=.220-volt.ru, backendEnd=1713814939465; expires=Sat, 21 Apr 2029 19:42:19 GMT; path=/; domain=.220-volt.ru, system_id=7fd0c2107a2bca39c71256e3cdb15a18z1713814939; path=/;Max-Age=1800;Domain=.220-volt.ru', 'Expires': 'Thu, 01 Jan 1970 00:00:00 GMT', 'Access-Control-Allow-Headers': 'Content-Type', 'Access-Control-Allow-Origin': 'https://*220-volt.ru', 'Cache-Control': 'no-store, no-cache, must-revalidate, max-age=0', 'X-Elapsed': '0.0526', 'Pragma': 'no-cache', 'Access-Control-Allow-Methods': 'OPTIONS, GET, POST, PUT, HEAD, PATCH, DELETE, OPTIONS', 'Last-Modified': 'Mon, 22 Apr 2024 19:42:19 GMT', 'RID': 'Zia9m9KYiT8da6FV-ARw5gAAABY', 'Content-Encoding': 'gzip'}
+Response cookies: {'chkuidsrv': '35e2ca3a69fa466aadf3bed420a32884', 'site_version': 'desktop', 'telemetryToken': 'A53WpsneRNDlDxriGk0KwP5wDuY6VIbuRi2EimNKzvuP', 'backendStart': '1713814939412', 'backendEnd': '1713814939465', 'system_id': '7fd0c2107a2bca39c71256e3cdb15a18z1713814939'}
+
+-----
