+++ tests/test_krasnodar_220_volt_api.py
@@ -1,6 +1,9 @@
 import allure
 from utils.checking import Checking
 from utils.api import Krasnodar220VoltApi
+from utils.pydantic_model import validate_data
+from utils.pydantic_model import Products
+
 
 """Получения информации о категории и продуктах"""
 
@@ -12,6 +15,11 @@ class TestTakeCategoryAndProductList():
         print("Метод POST")
         result_post = Krasnodar220VoltApi.category_and_product_list()
         Checking.check_status_code(result_post, 200)
+
+    def test_validate_data(self):
+        validate_data(Krasnodar220VoltApi.category_and_product_list().json(), Products)
+
+
